Fix a harmless compilation warning.
git-svn-id: http://yate.null.ro/svn/yate/trunk@4968 acf43c95-373e-0410-b603-e72c3f656dc1

+++ libs/yrtp/dejitter.cpp
@@ -176,7 +176,7 @@ void RTPDejitter::timerTick(const Time& when)
     unsigned int count = 0;
     while ((packet = static_cast<RTPDelayedData*>(m_packets.get()))) {
 	long delayed = when - packet->scheduled();
-	if (delayed <= 0 || delayed <= m_minDelay)
+	if (delayed <= 0 || delayed <= (long)m_minDelay)
 	    break;
 	// we are too delayed - probably rtpRecv() took too long to complete...
 	m_packets.remove(packet,true);
